refactor(web) : store의 import, export 컨벤션을 통일한다
