+++ packages/web/src/app/[lng]/(main)/community/[articleId]/components/ReplyList.tsx
@@ -1,7 +1,7 @@
 import { Reply } from '@/apis/community';
 import ReplyItem from '@/app/[lng]/(main)/community/[articleId]/components/ReplyItem';
 import { ItemList } from '@/components/List';
-import { useBlockStore } from '@/store/useBlockStore';
+import { useBlockStore } from '@/store';
 
 interface ReplyListProps {
   replyList: Reply[];
