+++ packages/web/src/app/[lng]/(main)/community/components/QuestionContent.client.tsx
@@ -8,7 +8,7 @@ import Empty from './Empty';
 
 import { useGetCommunityArticles } from '@/apis/community/queries';
 import { ItemList } from '@/components/List';
-import { useBlockStore } from '@/store/useBlockStore';
+import { useBlockStore } from '@/store';
 
 export default function QuestionContent() {
   const { ref, inView } = useInView();
