+++ packages/web/src/app/[lng]/(main)/grouping/[groupId]/articles/[articleId]/components/CommentList.client.tsx
@@ -9,7 +9,7 @@ import { Flex } from '@/components/Layout';
 import { ItemList } from '@/components/List';
 import { Spacing } from '@/components/Spacing';
 import { useNumberParams } from '@/hooks';
-import { useBlockStore } from '@/store/useBlockStore';
+import { useBlockStore } from '@/store';
 
 export default function CommentList() {
   const { articleId, groupId } = useNumberParams<['articleId', 'groupId']>();
