+++ packages/web/src/app/[lng]/(main)/grouping/[groupId]/components/articles/ArticleSection.client.tsx
@@ -5,7 +5,7 @@ import { useGetArticles } from '@/apis/groups/queries';
 import ArticleItem from '@/app/[lng]/(main)/grouping/components/ArticleItem.client';
 import { ItemList } from '@/components/List';
 import { useNumberParams } from '@/hooks';
-import { useBlockStore } from '@/store/useBlockStore';
+import { useBlockStore } from '@/store';
 
 interface ArticlesContentProps extends GroupDetailResponse {}
