+++ packages/web/src/app/[lng]/(main)/grouping/hooks/useMoreSheet.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from '@/app/i18n/client';
 import MoreBottomSheet from '@/components/Modal/MoreBottomSheet.client';
 import { useAppRouter } from '@/hooks';
 import { useModal } from '@/hooks';
-import { useBlockStore } from '@/store/useBlockStore';
+import { useBlockStore } from '@/store';
 
 type CommentId<T> = T extends 'comment' ? { commentId: number } : { commentId?: never };
