+++ packages/web/src/app/[lng]/(main)/meeting/participate/feedback/[groupId]/components/FeedbackFunnel.client.tsx
@@ -4,10 +4,10 @@ import FeedbackHeader from './FeedbackHeader';
 import Step1 from '../funnels/step1/Step1';
 import Step2 from '../funnels/step2/Step2';
 import Step3 from '../funnels/step3/Step3';
+
 import { useGetEstimate } from '@/apis/groups';
 import { useTranslation } from '@/app/i18n/client';
-import { useFunnel } from '@/hooks/useFunnel';
-import { useNumberParams } from '@/hooks';
+import { useFunnel, useNumberParams } from '@/hooks';
 
 export default function FeedbackSection() {
   const { Funnel, prevStep, nextStep } = useFunnel(['1', '2', '3']);
