+++ packages/web/src/hooks/useBrowser.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { useDidMount } from './common/useDidMount';
+import useDidMount from './useDidMount';
 
 export default function useBrowser() {
   const [browser, setBrowser] = useState('');
